requirements.txt
numpy
pillow
cryptography
opencv-python
